+++ src/internal/expr-interp.c
@@ -281,6 +281,12 @@ static sexp* deep_big_bang_coerce(sexp* x) {
     FREE(n_protect);
     return x;
   }
+  case r_type_s4: {
+    x = KEEP(r_eval_with_x(as_list_s4_call, r_methods_ns_env, x));
+    x = r_vec_coerce(x, r_type_pairlist);
+    FREE(1);
+    return x;
+  }
   case r_type_call:
     if (r_is_symbol(r_node_car(x), "{")) {
       return r_node_cdr(x);