+++ src/internal/internal.c
@@ -3,6 +3,7 @@
 
 
 sexp* as_list_call = NULL;
+sexp* as_list_s4_call = NULL;
 
 void rlang_init_dots();
 void rlang_init_eval_tidy();
@@ -13,6 +14,9 @@ void rlang_init_internal() {
 
   as_list_call = r_parse("as.list(x)");
   r_mark_precious(as_list_call);
+
+  as_list_s4_call = r_parse("as(x, 'list')");
+  r_mark_precious(as_list_s4_call);
 
   /* dots.c - enum dots_expansion_op */
   RLANG_ASSERT(OP_DOTS_MAX == DOTS_CAPTURE_TYPE_MAX * EXPANSION_OP_MAX);